pyproject.toml
[tool.poetry]
name = "start-stop-instances"
version = "0.2.0"
description = ""
authors = ["Abhinav Chaturvedi <achaturvedi@calance.com>"]
readme = "README.md"
packages = [{include = "start_stop_instances"}]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.9.0"
boto3 = "^1.28.25"
python-dotenv = "^1.0.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-stop-instances = "start_stop_instances.start_stop_instances:app"